+++ examples/demuxing.rs
@@ -2,15 +2,16 @@ use std::fs::File;
 
 use ac_ffmpeg::{
     format::{
-        demuxer::{Demuxer, DemuxerWithCodecParameters},
+        demuxer::{Demuxer, DemuxerWithStreamInfo, SeekTarget},
         io::IO,
     },
+    time::Timestamp,
     Error,
 };
 use clap::{App, Arg};
 
 /// Open a given input file.
-fn open_input(path: &str) -> Result<DemuxerWithCodecParameters<File>, Error> {
+fn open_input(path: &str) -> Result<DemuxerWithStreamInfo<File>, Error> {
     let input = File::open(path)
         .map_err(|err| Error::new(format!("unable to open input file {}: {}", path, err)))?;
 
@@ -26,8 +27,11 @@ fn open_input(path: &str) -> Result<DemuxerWithCodecParameters<File>, Error> {
 fn print_info(input: &str) -> Result<(), Error> {
     let mut demuxer = open_input(input)?;
 
-    for (index, params) in demuxer.codec_parameters().iter().enumerate() {
+    for (index, stream) in demuxer.streams().iter().enumerate() {
+        let params = stream.codec_parameters();
+
         println!("Stream #{}:", index);
+        println!("  duration: {}", stream.duration().as_f64().unwrap_or(0f64));
 
         if let Some(params) = params.as_audio_codec_parameters() {
             println!("  type: audio");
@@ -45,6 +49,9 @@ fn print_info(input: &str) -> Result<(), Error> {
             println!("  type: unknown");
         }
     }
+
+    println!("\nSeeking to START + 5s...");
+    demuxer.seek_to_timestamp(Timestamp::from_secs(5), SeekTarget::From)?;
 
     println!("\nPackets:");