+++ src/format/mod.rs
@@ -3,3 +3,4 @@
 pub mod demuxer;
 pub mod io;
 pub mod muxer;
+pub mod stream;